refactor(succinct): trim obsolete script crates
