+++ Cargo.toml
@@ -50,13 +50,11 @@ members = [
   "crates/proof/challenge",
   "crates/proof/succinct/utils/client",
   "crates/proof/succinct/utils/host",
-  "crates/proof/succinct/utils/build",
   "crates/proof/succinct/utils/proof",
-  "crates/proof/succinct/utils/signer",
   "crates/proof/succinct/utils/elfs",
   "crates/proof/succinct/utils/ethereum/client",
   "crates/proof/succinct/utils/ethereum/host",
-  "crates/proof/succinct/scripts/*",
+  "crates/proof/succinct/scripts",
 
   "crates/utilities/*",
   "crates/batcher/*",
@@ -267,13 +265,10 @@ base-proof-tee-nitro-attestation-prover = { path = "crates/proof/tee/nitro-attes
 
 # Succinct (OP Succinct / SP1 proving)
 base-proof-succinct-elfs = { path = "crates/proof/succinct/utils/elfs" }
-base-proof-succinct-prove = { path = "crates/proof/succinct/scripts/prove" }
-base-proof-succinct-scripts = { path = "crates/proof/succinct/scripts/utils" }
+base-proof-succinct-scripts = { path = "crates/proof/succinct/scripts" }
 base-proof-succinct-host-utils = { path = "crates/proof/succinct/utils/host" }
-base-proof-succinct-build-utils = { path = "crates/proof/succinct/utils/build" }
 base-proof-succinct-proof-utils = { path = "crates/proof/succinct/utils/proof" }
 base-proof-succinct-client-utils = { path = "crates/proof/succinct/utils/client" }
-base-proof-succinct-signer-utils = { path = "crates/proof/succinct/utils/signer" }
 base-proof-succinct-ethereum-host-utils = { path = "crates/proof/succinct/utils/ethereum/host" }
 base-proof-succinct-ethereum-client-utils = { path = "crates/proof/succinct/utils/ethereum/client" }
 
@@ -475,7 +470,6 @@ risc0-ethereum-contracts = "3.0.1"
 risc0-zkvm = { version = "^3.0", default-features = false }
 
 # SP1 v6.2.1 (Hypercube) — used by crates/proof/succinct
-sp1-build = "=6.2.1"
 sp1-prover-types = "=6.2.1"
 sp1-sdk = { version = "=6.2.1" }
 sp1-cluster-utils = { git = "https://github.com/succinctlabs/sp1-cluster", tag = "v2.3.2" }
