+++ crates/proof/succinct/scripts/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "base-proof-succinct-scripts"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[[bin]]
+name = "vkeys"
+path = "bin/vkeys.rs"
+
+[[bin]]
+name = "parse-receipt"
+path = "bin/parse_receipt.rs"
+
+[dependencies]
+
+# workspace
+clap.workspace = true
+tokio.workspace = true
+anyhow.workspace = true
+bincode.workspace = true
+serde_json.workspace = true
+alloy-primitives.workspace = true
+
+# local
+base-proof-succinct-elfs.workspace = true
+base-proof-succinct-proof-utils.workspace = true
+base-proof-succinct-client-utils.workspace = true
+
+# sp1
+sp1-sdk = { workspace = true, features = ["blocking"] }
+
+[lints]
+workspace = true
