+++ crates/proof/succinct/scripts/bin/parse_receipt.rs
