--- crates/proof/succinct/scripts/prove/Cargo.toml
@@ -1,66 +1,0 @@
-[package]
-name = "base-proof-succinct-prove"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[[bin]]
-name = "multi"
-path = "bin/multi.rs"
-
-[[bin]]
-name = "agg"
-path = "bin/agg.rs"
-
-[dependencies]
-
-# workspace
-tokio.workspace = true
-clap.workspace = true
-anyhow.workspace = true
-dotenv.workspace = true
-csv.workspace = true
-tracing.workspace = true
-cfg-if.workspace = true
-serde_json.workspace = true
-
-# local
-base-proof-succinct-host-utils.workspace = true
-base-proof-succinct-client-utils.workspace = true
-base-proof-succinct-proof-utils.workspace = true
-base-proof-succinct-scripts = { path = "../utils" }
-base-proof-succinct-elfs.workspace = true
-
-# alloy
-alloy-contract.workspace = true
-alloy-eips.workspace = true
-alloy-network.workspace = true
-alloy-primitives.workspace = true
-alloy-provider = { workspace = true, features = ["reqwest"] }
-alloy-signer-local.workspace = true
-alloy-sol-types.workspace = true
-alloy-transport-http = { workspace = true, features = [
-    "reqwest",
-    "reqwest-native-tls",
-] }
-
-# sp1
-sp1-sdk = { workspace = true, features = ["network", "blocking", "profiling"] }
-sp1-cluster-utils.workspace = true
-
-[dev-dependencies]
-reqwest = { version = "0.12.4", features = ["json"] }
-rustls = { version = "0.23", features = ["ring"] }
-base-proof-preimage.workspace = true
-tempfile.workspace = true
-gag = "1.0"
-
-[build-dependencies]
-base-proof-succinct-build-utils.workspace = true
-
-[lints]
-workspace = true
-
