--- crates/proof/succinct/scripts/prove/build.rs
@@ -1,8 +1,0 @@
-//! Build script for the prove scripts.
-
-use base_proof_succinct_build_utils::build_all;
-
-/// Build all the prove scripts.
-fn main() {
-    build_all();
-}
