--- crates/proof/succinct/scripts/prove/src/lib.rs
@@ -1,41 +1,0 @@
-//! Proving utilities for executing multi-block range proofs.
-
-use std::time::{Duration, Instant};
-
-use anyhow::{Ok, Result};
-use base_proof_succinct_host_utils::fetcher::{BlockInfo, OPSuccinctDataFetcher};
-use base_proof_succinct_proof_utils::get_range_elf_embedded;
-use sp1_sdk::{
-    Elf, ExecutionReport, SP1Stdin,
-    blocking::{CpuProver, Prover},
-};
-
-/// Duration representing two weeks, used as a default proving window.
-pub const TWO_WEEKS: Duration = Duration::from_secs(14 * 24 * 60 * 60);
-
-/// Executes a multi-block range proof and returns block data, execution report, and duration.
-pub async fn execute_multi(
-    data_fetcher: &OPSuccinctDataFetcher,
-    sp1_stdin: SP1Stdin,
-    l2_start_block: u64,
-    l2_end_block: u64,
-) -> Result<(Vec<BlockInfo>, ExecutionReport, Duration)> {
-    let start_time = Instant::now();
-
-    // CpuProver creates its own tokio runtime, so run it outside the async context.
-    let (_, report) = tokio::task::spawn_blocking(move || {
-        let prover = CpuProver::new();
-        prover
-            .execute(Elf::Static(get_range_elf_embedded()), sp1_stdin)
-            .calculate_gas(true)
-            .deferred_proof_verification(false)
-            .run()
-    })
-    .await??;
-
-    let execution_duration = start_time.elapsed();
-
-    let block_data = data_fetcher.get_l2_block_data_range(l2_start_block, l2_end_block).await?;
-
-    Ok((block_data, report, execution_duration))
-}
