--- crates/proof/succinct/scripts/prove/tests/exploit_regression.rs
@@ -1,225 +1,0 @@
-//! Real exploit-regression test for GHSA-5jh4-3p33-85xc.
-//!
-//! Generates an honest witness for a small L2 range ending at block N. The mutation step then
-//! injects the GHSA exploit shape on top of the honest witness:
-//!
-//!   - `L1_HEAD_KEY` (local 1) → the L1 block where N becomes safe (returned by
-//!     `data_fetcher.get_safe_l1_block_for_l2_block(N)`). This tightens the L1 chain provider's
-//!     head from the honest `+20` buffer down to the channel-close boundary, so the pipeline runs
-//!     out of L1 origin before it can derive past N.
-//!   - `L2_CLAIM_BLOCK_NUMBER_KEY` (local 4) → `T = N + 100`. The witness now claims a target the
-//!     data cannot reach.
-//!
-//! `claimed_l2_output_root = root(N)` and the rest of the witness are left intact.
-//!
-//! The test:
-//!   1. Captures host stderr to a tempfile via `gag::Redirect::stderr`.
-//!   2. Asserts `report.exit_code != 0` (the guest panicked).
-//!   3. Rejects the run if the captured stderr contains `InvalidPreimageKey` or `Trie walker
-//!      error` — those mean derivation drifted past N and hit a non-postcondition failure.
-//!   4. Parses the derived safe-head block number and the claimed block number from the captured
-//!      panic message. Asserts the claimed value equals the mutated `T` exactly AND the derived
-//!      value equals the honest `N`.
-//!
-//! Manual / release-verification only. Ignored by default. Run explicitly with:
-//!
-//!   `cargo test -p base-proof-succinct-prove --test exploit_regression -- --ignored --nocapture`
-//!
-//! Requires live RPC env (`L1_RPC`, `L2_RPC`, `L2_NODE_RPC`).
-//!
-//! Gated to `unix` because `gag::Redirect::stderr` is Unix-only.
-
-#![cfg(unix)]
-
-use std::{
-    env,
-    io::{Read, Seek, SeekFrom},
-    sync::Arc,
-};
-
-use anyhow::Result;
-use base_proof_preimage::PreimageKey;
-use base_proof_succinct_client_utils::client::DEFAULT_INTERMEDIATE_ROOT_INTERVAL;
-use base_proof_succinct_host_utils::{
-    block_range::get_rolling_block_range, fetcher::OPSuccinctDataFetcher, host::OPSuccinctHost,
-    witness_generation::WitnessGenerator,
-};
-use base_proof_succinct_proof_utils::{get_range_elf_embedded, initialize_host};
-use gag::Redirect;
-use sp1_sdk::{
-    Elf, SP1Stdin,
-    blocking::{CpuProver, Prover},
-};
-
-const L1_HEAD_LOCAL_IDENT: u64 = 1;
-const L2_CLAIM_BLOCK_NUMBER_LOCAL_IDENT: u64 = 4;
-const ADVERSARIAL_OFFSET: u64 = 100;
-const REQUIRED_ENV_VARS: &[&str] = &["L1_RPC", "L2_RPC", "L2_NODE_RPC"];
-
-const POSTCOND_DERIVED_LABEL: &str = "Derived safe head L2 block #";
-const POSTCOND_CLAIMED_LABEL: &str = "claimed L2 block number #";
-const NON_POSTCOND_NEEDLES: &[&str] = &["Invalid preimage key", "Trie walker error"];
-
-fn missing_required_env() -> Vec<&'static str> {
-    REQUIRED_ENV_VARS
-        .iter()
-        .copied()
-        .filter(|var| env::var(var).map(|v| v.is_empty()).unwrap_or(true))
-        .collect()
-}
-
-fn parse_number_after(haystack: &str, marker: &str) -> Option<u64> {
-    let start = haystack.find(marker)? + marker.len();
-    let rest = &haystack[start..];
-    let end = rest.find(|c: char| !c.is_ascii_digit()).unwrap_or(rest.len());
-    if end == 0 { None } else { rest[..end].parse().ok() }
-}
-
-#[ignore = "opt-in exploit regression: set L1_RPC, L2_RPC, L2_NODE_RPC env vars"]
-#[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-async fn ghsa_5jh4_3p33_85xc_silent_downgrade_rejected_by_real_range_elf() -> Result<()> {
-    dotenv::dotenv().ok();
-
-    let missing = missing_required_env();
-    if !missing.is_empty() {
-        anyhow::bail!(
-            "missing required env vars {missing:?}; this test drives an honest witness through \
-             the real, embedded range ELF and requires live RPC access"
-        );
-    }
-
-    let data_fetcher = OPSuccinctDataFetcher::new_with_rollup_config().await?;
-    let host = initialize_host(Arc::new(data_fetcher.clone()));
-
-    const RANGE: u64 = 2;
-    let (l2_start_block, honest_n) =
-        get_rolling_block_range(host.as_ref(), &data_fetcher, RANGE).await?;
-    let claimed_t = honest_n.checked_add(ADVERSARIAL_OFFSET).expect("T fits in u64");
-
-    eprintln!(
-        "ghsa-5jh4-3p33-85xc: honest range = [{l2_start_block}, {honest_n}]; \
-         adversarial T = {claimed_t} (= N + {ADVERSARIAL_OFFSET})"
-    );
-
-    // Generate honest witness for [S, N].
-    let host_args = host
-        .fetch(l2_start_block, honest_n, None, DEFAULT_INTERMEDIATE_ROOT_INTERVAL, false)
-        .await?;
-    let mut witness_data = host.run(&host_args).await?;
-
-    // Sanity: honest witness commits N before mutation.
-    let claim_block_key = PreimageKey::new_local(L2_CLAIM_BLOCK_NUMBER_LOCAL_IDENT);
-    let honest_claim_bytes = witness_data
-        .preimage_store
-        .preimage_map
-        .get(&claim_block_key)
-        .expect("honest witness must contain L2_CLAIM_BLOCK_NUMBER_KEY")
-        .clone();
-    let honest_claim_block = u64::from_be_bytes(
-        <[u8; 8]>::try_from(honest_claim_bytes.as_slice())
-            .expect("L2_CLAIM_BLOCK_NUMBER_KEY value must be u64 BE"),
-    );
-    assert_eq!(honest_claim_block, honest_n, "honest witness should claim range upper bound");
-
-    // Mutation #1: tighten L1_HEAD to the L1 block where N becomes safe.
-    let (tight_l1_head_hash, tight_l1_head_number) =
-        data_fetcher.get_safe_l1_block_for_l2_block(honest_n).await?;
-    eprintln!(
-        "ghsa-5jh4-3p33-85xc: tightening L1_HEAD to L1 block #{tight_l1_head_number} \
-         hash {tight_l1_head_hash:?}"
-    );
-    let l1_head_key = PreimageKey::new_local(L1_HEAD_LOCAL_IDENT);
-    witness_data.preimage_store.preimage_map.insert(l1_head_key, tight_l1_head_hash.0.to_vec());
-
-    // Mutation #2: claim a block T = N + 100 the data cannot reach.
-    witness_data
-        .preimage_store
-        .preimage_map
-        .insert(claim_block_key, claimed_t.to_be_bytes().to_vec());
-
-    let sp1_stdin: SP1Stdin = host.witness_generator().get_sp1_stdin(witness_data)?;
-
-    // Capture stderr to a tempfile — the SP1 simulator forwards guest panics to host stderr.
-    let mut stderr_file = tempfile::tempfile()?;
-    let redirect_handle = stderr_file.try_clone()?;
-
-    let join_result = {
-        let _redirect = Redirect::stderr(redirect_handle).expect("redirect stderr to tempfile");
-        tokio::task::spawn_blocking(move || {
-            CpuProver::new()
-                .execute(Elf::Static(get_range_elf_embedded()), sp1_stdin)
-                .deferred_proof_verification(false)
-                .run()
-        })
-        .await
-    };
-
-    stderr_file.seek(SeekFrom::Start(0))?;
-    let mut captured_stderr = String::new();
-    stderr_file.read_to_string(&mut captured_stderr)?;
-    eprintln!("--- captured guest stderr ---\n{captured_stderr}\n--- end captured stderr ---");
-
-    let (pv, report) = match join_result? {
-        Ok(ok) => ok,
-        Err(host_err) => panic!(
-            "host-side execute() failed before the guest could run; postcondition was never \
-             reached. error: {host_err:?}\ncaptured stderr:\n{captured_stderr}"
-        ),
-    };
-
-    if report.exit_code == 0 {
-        panic!(
-            "patched range ELF returned exit_code = 0 — the (l2_output_root, l2_block_number) \
-             binding may have regressed.\n\
-             public_values len = {}\n\
-             captured stderr:\n{captured_stderr}",
-            pv.as_slice().len(),
-        );
-    }
-
-    for needle in NON_POSTCOND_NEEDLES {
-        if captured_stderr.contains(needle) {
-            panic!(
-                "guest panicked, but with a non-postcondition error containing {needle:?}.\n\
-                 The two-key mutation strategy did not steer derivation to a clean EndOfSource \
-                 path; iterate on the test (e.g., adjust the range / tight head / RANGE size).\n\
-                 captured stderr:\n{captured_stderr}",
-            );
-        }
-    }
-
-    assert!(
-        captured_stderr.contains(POSTCOND_DERIVED_LABEL)
-            && captured_stderr.contains(POSTCOND_CLAIMED_LABEL),
-        "guest panicked with non-zero exit_code but the captured stderr does not contain \
-         the postcondition mismatch labels.\n\
-         expected: contains {POSTCOND_DERIVED_LABEL:?} AND {POSTCOND_CLAIMED_LABEL:?}\n\
-         captured stderr:\n{captured_stderr}"
-    );
-
-    let derived_in_err = parse_number_after(&captured_stderr, POSTCOND_DERIVED_LABEL).expect(
-        "postcondition error must include the derived safe-head block number after its label",
-    );
-    let claimed_in_err = parse_number_after(&captured_stderr, POSTCOND_CLAIMED_LABEL)
-        .expect("postcondition error must include the claimed block number after its label");
-
-    assert_eq!(
-        claimed_in_err, claimed_t,
-        "error's claimed L2 block number ({claimed_in_err}) must equal the mutated T ({claimed_t})"
-    );
-    assert_eq!(
-        derived_in_err, honest_n,
-        "GHSA invariant: derived safe head must equal the honest N. Only when derivation \
-         halts at exactly N does the silent-downgrade output root match the unmutated \
-         claimed_l2_output_root = root(N) — i.e., the postcondition is what rejects. \
-         derived = {derived_in_err}, honest N = {honest_n}, T = {claimed_t}",
-    );
-
-    eprintln!(
-        "ghsa-5jh4-3p33-85xc: postcondition rejected exploit. \
-         derived = {derived_in_err}, claimed T = {claimed_in_err}, exit_code = {}.",
-        report.exit_code,
-    );
-
-    Ok(())
-}
