--- crates/proof/succinct/scripts/utils/Cargo.toml
@@ -1,83 +1,0 @@
-[package]
-name = "base-proof-succinct-scripts"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[lib]
-name = "base_proof_succinct_scripts"
-path = "src/lib.rs"
-
-[[bin]]
-name = "fetch-and-save-proof"
-path = "bin/fetch_and_save_proof.rs"
-
-[[bin]]
-name = "config"
-path = "bin/config.rs"
-
-[[bin]]
-name = "cost-estimator"
-path = "bin/cost_estimator.rs"
-
-[[bin]]
-name = "block-data"
-path = "bin/block_data.rs"
-
-[[bin]]
-name = "fetch-l2oo-config"
-path = "bin/fetch_l2oo_config.rs"
-
-[[bin]]
-name = "gen-sp1-test-artifacts"
-path = "bin/gen_sp1_test_artifacts.rs"
-
-[[bin]]
-name = "parse-receipt"
-path = "bin/parse_receipt.rs"
-
-[dependencies]
-
-# workspace
-alloy-network.workspace = true
-alloy-provider = { workspace = true, features = ["reqwest"] }
-alloy-transport-http = { workspace = true, features = [
-    "reqwest",
-    "reqwest-native-tls",
-] }
-alloy-primitives.workspace = true
-alloy-sol-types.workspace = true
-alloy-signer-local.workspace = true
-alloy-eips.workspace = true
-bincode.workspace = true
-tokio.workspace = true
-clap.workspace = true
-cargo_metadata.workspace = true
-anyhow.workspace = true
-dotenv.workspace = true
-log.workspace = true
-csv.workspace = true
-serde.workspace = true
-futures.workspace = true
-rayon = "1.10.0"
-serde_json.workspace = true
-tracing.workspace = true
-
-# local
-base-proof-succinct-host-utils.workspace = true
-base-proof-succinct-client-utils.workspace = true
-base-proof-succinct-proof-utils.workspace = true
-base-proof-succinct-elfs.workspace = true
-
-# sp1
-sp1-sdk = { workspace = true, features = ["network", "blocking", "profiling"] }
-
-[build-dependencies]
-base-proof-succinct-build-utils.workspace = true
-
-[lints]
-workspace = true
-
