--- crates/proof/succinct/scripts/utils/build.rs
@@ -1,7 +1,0 @@
-//! Build script for the succinct scripts utility crate.
-
-use base_proof_succinct_build_utils::build_all;
-
-fn main() {
-    build_all();
-}
