--- crates/proof/succinct/scripts/utils/src/config_common.rs
@@ -1,132 +1,0 @@
-use std::{
-    env, fs,
-    path::{Path, PathBuf},
-};
-
-use alloy_primitives::{Address, hex};
-use alloy_signer_local::PrivateKeySigner;
-use anyhow::Result;
-use base_proof_succinct_client_utils::{boot::hash_rollup_config, types::u32_to_u8};
-use base_proof_succinct_elfs::AGGREGATION_ELF;
-use base_proof_succinct_host_utils::fetcher::OPSuccinctDataFetcher;
-use base_proof_succinct_proof_utils::get_range_elf_embedded;
-use sp1_sdk::{Elf, HashableKey, Prover, ProverClient, ProvingKey};
-
-/// Two weeks expressed in seconds.
-pub const TWO_WEEKS_IN_SECONDS: u64 = 14 * 24 * 60 * 60;
-
-/// Shared configuration data that both L2OO and FDG configs use.
-#[derive(Debug, Clone)]
-pub struct SharedConfigData {
-    /// Hash of the rollup configuration.
-    pub rollup_config_hash: String,
-    /// Aggregation verification key.
-    pub aggregation_vkey: String,
-    /// Commitment to the range verification key.
-    pub range_vkey_commitment: String,
-    /// On-chain verifier contract address.
-    pub verifier_address: String,
-    /// Whether to use the SP1 mock verifier instead of real verification.
-    pub use_sp1_mock_verifier: bool,
-}
-
-/// Returns an address based on environment variables and private key settings:
-/// - If `env_var` exists, returns that address
-/// - Otherwise if `private_key_by_default=true` and `PRIVATE_KEY` exists, returns address derived from
-///   private key
-/// - Otherwise returns zero address
-pub fn get_address(env_var: &str, private_key_by_default: bool) -> String {
-    // First try to get address directly from env var.
-    if let Ok(addr) = env::var(env_var) {
-        return addr;
-    }
-
-    // Next try to derive address from private key if enabled.
-    if private_key_by_default && let Ok(pk) = env::var("PRIVATE_KEY") {
-        let signer: PrivateKeySigner = pk.parse().unwrap();
-        return signer.address().to_string();
-    }
-
-    // Fallback to zero address.
-    Address::ZERO.to_string()
-}
-
-/// Parse comma-separated addresses from environment variable.
-pub fn parse_addresses(env_var: &str) -> Vec<String> {
-    env::var(env_var)
-        .unwrap_or_default()
-        .split(',')
-        .filter_map(|s| {
-            let trimmed = s.trim();
-            if trimmed.is_empty() { None } else { Some(trimmed.to_string()) }
-        })
-        .collect()
-}
-
-/// Get shared configuration data that both L2OO and FDG configs need.
-pub async fn get_shared_config_data(
-    data_fetcher: OPSuccinctDataFetcher,
-) -> Result<SharedConfigData> {
-    // Determine if we're using mock verifier.
-    let use_sp1_mock_verifier = env::var("OP_SUCCINCT_MOCK")
-        .unwrap_or_else(|_| "false".to_string())
-        .parse::<bool>()
-        .unwrap_or(false);
-
-    // Set the verifier address.
-    let verifier_address = env::var("VERIFIER_ADDRESS").unwrap_or_else(|_| {
-        // Default to Plonk VerifierGateway contract address.
-        // Source: https://docs.succinct.xyz/docs/sp1/verification/contract-addresses
-        "0x3B6041173B80E77f038f3F2C0f9744f04837185e".to_string()
-    });
-
-    let rollup_config = data_fetcher.rollup_config.as_ref().unwrap();
-    let rollup_config_hash = format!("0x{:x}", hash_rollup_config(rollup_config));
-
-    // Calculate verification keys.
-    let prover = ProverClient::builder().cpu().build().await;
-    let agg_pk = prover.setup(Elf::Static(AGGREGATION_ELF)).await?;
-    let agg_vkey = agg_pk.verifying_key();
-    let aggregation_vkey = agg_vkey.bytes32();
-
-    let range_pk = prover.setup(Elf::Static(get_range_elf_embedded())).await?;
-    let range_vkey = range_pk.verifying_key();
-    let range_vkey_commitment = format!("0x{}", hex::encode(u32_to_u8(range_vkey.hash_u32())));
-
-    Ok(SharedConfigData {
-        rollup_config_hash,
-        aggregation_vkey,
-        range_vkey_commitment,
-        verifier_address,
-        use_sp1_mock_verifier,
-    })
-}
-
-/// Write a JSON config to a file.
-pub fn write_config_file<T: serde::Serialize>(
-    config: &T,
-    file_path: &Path,
-    description: &str,
-) -> Result<()> {
-    // Create parent directories if they don't exist.
-    if let Some(parent) = file_path.parent() {
-        fs::create_dir_all(parent)?;
-    }
-    // Write the config to the file.
-    fs::write(file_path, serde_json::to_string_pretty(config)?)?;
-
-    log::info!("Wrote {} configuration to: {}", description, file_path.display());
-
-    Ok(())
-}
-
-/// Find the project root directory (where .git exists).
-pub fn find_project_root() -> Option<PathBuf> {
-    let mut path = std::env::current_dir().ok()?;
-    while !path.join(".git").exists() {
-        if !path.pop() {
-            return None;
-        }
-    }
-    Some(path)
-}
