--- crates/proof/succinct/scripts/utils/src/lib.rs
@@ -1,63 +1,0 @@
-//! Shared CLI argument types and configuration helpers for succinct scripts.
-
-use std::path::PathBuf;
-
-use clap::Parser;
-
-/// Common configuration helpers shared across scripts.
-pub mod config_common;
-
-/// The arguments for the host executable.
-#[derive(Debug, Clone, Parser)]
-pub struct HostExecutorArgs {
-    /// The start block of the range to execute.
-    #[arg(long)]
-    pub start: Option<u64>,
-    /// The end block of the range to execute.
-    #[arg(long)]
-    pub end: Option<u64>,
-    /// The number of blocks to execute in a single batch.
-    #[arg(long, default_value = "10")]
-    pub batch_size: u64,
-    /// Enable caching: load from cache if available, save to cache if not.
-    #[arg(long)]
-    pub cache: bool,
-    /// Use a fixed recent range.
-    #[arg(long)]
-    pub rolling: bool,
-    /// The number of blocks to use for the default range.
-    #[arg(long, default_value = "5")]
-    pub default_range: u64,
-    /// The environment file to use.
-    #[arg(long, default_value = ".env")]
-    pub env_file: PathBuf,
-    /// Whether to generate proofs.
-    #[arg(long)]
-    pub prove: bool,
-    /// Whether to fallback to timestamp-based L1 head estimation even though `SafeDB` is not
-    /// activated for op-node.
-    #[clap(long)]
-    pub safe_db_fallback: bool,
-    /// Cluster proving timeout in seconds (only used when `SP1_PROVER=cluster`).
-    #[arg(long, default_value = "21600")]
-    pub cluster_timeout: u64,
-}
-
-impl HostExecutorArgs {
-    /// When both start and end are explicitly provided, default the batch size
-    /// to the full range so it is processed as a single batch.
-    pub const fn effective_batch_size(&self) -> u64 {
-        match (self.start, self.end) {
-            (Some(start), Some(end)) if end > start => end - start,
-            _ => self.batch_size,
-        }
-    }
-}
-
-/// Arguments for configuration generation scripts.
-#[derive(Debug, Clone, Parser)]
-pub struct ConfigArgs {
-    /// The environment file to use.
-    #[arg(long)]
-    pub env_file: Option<PathBuf>,
-}
