--- crates/proof/succinct/utils/signer/Cargo.toml
@@ -1,37 +1,0 @@
-[package]
-name = "base-proof-succinct-signer-utils"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[dependencies]
-
-# alloy
-alloy-consensus.workspace = true
-alloy-eips.workspace = true
-alloy-network.workspace = true
-alloy-primitives.workspace = true
-alloy-provider.workspace = true
-alloy-rpc-types-eth.workspace = true
-alloy-signer.workspace = true
-alloy-signer-local.workspace = true
-alloy-signer-gcp.workspace = true
-alloy-transport-http.workspace = true
-
-# general
-anyhow.workspace = true
-tokio.workspace = true
-rustls = { version = "0.23", default-features = false, features = [
-    "std",
-    "tls12",
-    "aws-lc-rs",
-] }
-[dev-dependencies]
-dotenv.workspace = true
-base-proof-succinct-host-utils.workspace = true
-
-[lints]
-workspace = true
