--- crates/proof/succinct/utils/signer/src/lib.rs
@@ -1,303 +1,0 @@
-//! Transaction signing utilities supporting `Web3Signer`, local private keys, and Google Cloud HSM.
-
-use std::{str::FromStr, sync::Arc};
-
-use alloy_consensus::TxEnvelope;
-use alloy_eips::Decodable2718;
-use alloy_network::{Ethereum, EthereumWallet, TransactionBuilder};
-use alloy_primitives::{Address, Bytes, TxKind};
-use alloy_provider::{Provider, ProviderBuilder, Web3Signer};
-use alloy_rpc_types_eth::{TransactionReceipt, TransactionRequest};
-use alloy_signer::Signer as AlloySigner;
-use alloy_signer_gcp::{
-    GcpKeyRingRef, GcpSigner, KeySpecifier,
-    gcloud_sdk::{
-        GoogleApi,
-        google::cloud::kms::v1::key_management_service_client::KeyManagementServiceClient,
-    },
-};
-use alloy_signer_local::PrivateKeySigner;
-use alloy_transport_http::reqwest::Url;
-use anyhow::{Context, Result};
-use tokio::{sync::Mutex, time::Duration};
-
-/// Number of block confirmations to wait for before considering a transaction final.
-pub const NUM_CONFIRMATIONS: u64 = 3;
-/// Timeout in seconds for waiting on transaction receipts.
-pub const TIMEOUT_SECONDS: u64 = 60;
-
-#[derive(Clone, Debug)]
-/// The type of signer to use for signing transactions.
-pub enum Signer {
-    /// The signer URL and address.
-    Web3Signer(Url, Address),
-    /// The local signer.
-    LocalSigner(PrivateKeySigner),
-    /// Cloud HSM signer using Google.
-    CloudHsmSigner(GcpSigner),
-}
-
-impl Signer {
-    /// Returns the address associated with this signer.
-    pub fn address(&self) -> Address {
-        match self {
-            Self::Web3Signer(_, address) => *address,
-            Self::LocalSigner(signer) => signer.address(),
-            Self::CloudHsmSigner(signer) => signer.address(),
-        }
-    }
-
-    /// Creates a new Web3 signer with the given URL and address.
-    pub const fn new_web3_signer(url: Url, address: Address) -> Self {
-        Self::Web3Signer(url, address)
-    }
-
-    /// Creates a new local signer from a private key string.
-    pub fn new_local_signer(private_key_str: &str) -> Result<Self> {
-        let private_key =
-            PrivateKeySigner::from_str(private_key_str).context("Failed to parse private key")?;
-        Ok(Self::LocalSigner(private_key))
-    }
-
-    /// Creates a signer from environment variables.
-    ///
-    /// Checks for Cloud HSM, `Web3Signer`, and local private key configurations in that order.
-    pub async fn from_env() -> Result<Self> {
-        if let (Ok(project_id), Ok(location), Ok(keyring_name)) = (
-            std::env::var("GOOGLE_PROJECT_ID"),
-            std::env::var("GOOGLE_LOCATION"),
-            std::env::var("GOOGLE_KEYRING"),
-        ) {
-            let key_name = std::env::var("HSM_KEY_NAME").expect("HSM_KEY_NAME");
-            let key_version =
-                std::env::var("HSM_KEY_VERSION").unwrap_or_else(|_| "1".to_string()).parse()?;
-
-            let keyring = GcpKeyRingRef::new(&project_id, &location, &keyring_name);
-
-            let key_specifier = KeySpecifier::new(keyring, &key_name, key_version);
-
-            let client = GoogleApi::from_function(
-                KeyManagementServiceClient::new,
-                "https://cloudkms.googleapis.com",
-                None,
-            )
-            .await?;
-            let signer = GcpSigner::new(client, key_specifier, None).await?;
-
-            Ok(Self::CloudHsmSigner(signer))
-        } else if let (Ok(signer_url_str), Ok(signer_address_str)) =
-            (std::env::var("SIGNER_URL"), std::env::var("SIGNER_ADDRESS"))
-        {
-            let signer_url = Url::parse(&signer_url_str).context("Failed to parse SIGNER_URL")?;
-            let signer_address =
-                Address::from_str(&signer_address_str).context("Failed to parse SIGNER_ADDRESS")?;
-            Ok(Self::new_web3_signer(signer_url, signer_address))
-        } else if let Ok(private_key_str) = std::env::var("PRIVATE_KEY") {
-            Self::new_local_signer(&private_key_str)
-        } else {
-            anyhow::bail!(
-                "None of the required signer configurations are set in environment:\n\
-                - For Cloud HSM: GOOGLE_PROJECT_ID, GOOGLE_LOCATION, GOOGLE_KEYRING\n\
-                - For Web3Signer: SIGNER_URL and SIGNER_ADDRESS\n\
-                - For Local: PRIVATE_KEY"
-            )
-        }
-    }
-
-    /// Sends a transaction request, signed by the configured `signer`.
-    pub async fn send_transaction_request(
-        &self,
-        l1_rpc: Url,
-        mut transaction_request: TransactionRequest,
-    ) -> Result<TransactionReceipt> {
-        match self {
-            Self::Web3Signer(signer_url, signer_address) => {
-                // Set the from address to the signer address.
-                transaction_request.set_from(*signer_address);
-
-                // Fill the transaction request with all of the relevant gas and nonce information.
-                let provider = ProviderBuilder::new().network::<Ethereum>().connect_http(l1_rpc);
-                let filled_tx = provider.fill(transaction_request).await?;
-
-                // Sign the transaction request using the Web3Signer.
-                let web3_provider =
-                    ProviderBuilder::new().network::<Ethereum>().connect_http(signer_url.clone());
-                let signer = Web3Signer::new(web3_provider.clone(), *signer_address);
-
-                let mut tx = filled_tx.as_builder().unwrap().clone();
-                tx.normalize_data();
-
-                let raw: Bytes =
-                    signer.provider().client().request("eth_signTransaction", (tx,)).await?;
-
-                let tx_envelope = TxEnvelope::decode_2718(&mut raw.as_ref()).unwrap();
-
-                let receipt = provider
-                    .send_tx_envelope(tx_envelope)
-                    .await
-                    .context("Failed to send transaction")?
-                    .with_required_confirmations(NUM_CONFIRMATIONS)
-                    .with_timeout(Some(Duration::from_secs(TIMEOUT_SECONDS)))
-                    .get_receipt()
-                    .await?;
-
-                Ok(receipt)
-            }
-            Self::LocalSigner(private_key) => {
-                let provider = ProviderBuilder::new()
-                    .network::<Ethereum>()
-                    .wallet(EthereumWallet::new(private_key.clone()))
-                    .connect_http(l1_rpc);
-
-                // Ensure the request has a `from` address so the wallet filler can sign it.
-                transaction_request.set_from(private_key.address());
-                if transaction_request.to.is_none() {
-                    // NOTE(fakedev9999): Anvil's wallet filler insists on a `to` field even for
-                    // deployments. Mark the request as contract creation so it can be signed.
-                    transaction_request.to = Some(TxKind::Create);
-                }
-
-                let receipt = provider
-                    .send_transaction(transaction_request)
-                    .await
-                    .context("Failed to send transaction")?
-                    .with_required_confirmations(NUM_CONFIRMATIONS)
-                    .with_timeout(Some(Duration::from_secs(TIMEOUT_SECONDS)))
-                    .get_receipt()
-                    .await?;
-
-                Ok(receipt)
-            }
-            Self::CloudHsmSigner(signer) => {
-                // Set the from address to HSM address
-                transaction_request.set_from(signer.address());
-                if transaction_request.to.is_none() {
-                    // NOTE(fakedev9999): Anvil's wallet filler insists on a `to` field even for
-                    // deployments. Mark the request as contract creation so it can be signed.
-                    transaction_request.to = Some(TxKind::Create);
-                }
-
-                let wallet = EthereumWallet::new(signer.clone());
-                let provider = ProviderBuilder::new()
-                    .network::<Ethereum>()
-                    .wallet(wallet)
-                    .connect_http(l1_rpc);
-
-                let receipt = provider
-                    .send_transaction(transaction_request)
-                    .await
-                    .context("Failed to send KMS-signed transaction")?
-                    .with_required_confirmations(NUM_CONFIRMATIONS)
-                    .with_timeout(Some(Duration::from_secs(TIMEOUT_SECONDS)))
-                    .get_receipt()
-                    .await?;
-
-                Ok(receipt)
-            }
-        }
-    }
-}
-
-/// Wrapper around Signer that provides thread-safe transaction sending.
-/// Transactions are serialized via a Mutex to prevent nonce conflicts.
-#[derive(Clone, Debug)]
-pub struct SignerLock {
-    inner: Arc<Mutex<Signer>>,
-    cached_address: Address,
-}
-
-impl SignerLock {
-    /// Creates a new `SignerLock` wrapping the given Signer.
-    pub fn new(signer: Signer) -> Self {
-        let cached_address = signer.address();
-        Self { inner: Arc::new(Mutex::new(signer)), cached_address }
-    }
-
-    /// Creates a `SignerLock` from environment variables.
-    pub async fn from_env() -> Result<Self> {
-        Ok(Self::new(Signer::from_env().await?))
-    }
-
-    /// Returns the address of the signer without acquiring a lock.
-    pub const fn address(&self) -> Address {
-        self.cached_address
-    }
-
-    /// Sends a transaction request, signed by the configured signer.
-    /// Transactions are serialized via a Mutex to prevent nonce conflicts.
-    pub async fn send_transaction_request(
-        &self,
-        l1_rpc: Url,
-        transaction_request: TransactionRequest,
-    ) -> Result<TransactionReceipt> {
-        let signer = self.inner.lock().await;
-        signer.send_transaction_request(l1_rpc, transaction_request).await
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use alloy_eips::BlockId;
-    use alloy_primitives::{U256, address};
-    use base_proof_succinct_host_utils::OPSuccinctL2OutputOracle::OPSuccinctL2OutputOracleInstance as OPSuccinctL2OOContract;
-
-    use super::*;
-
-    #[tokio::test]
-    #[ignore]
-    async fn test_send_transaction_request_web3() {
-        let proposer_signer = SignerLock::new(Signer::new_web3_signer(
-            "http://localhost:9000".parse().unwrap(),
-            "0x9b3F173823E944d183D532ed236Ee3B83Ef15E1d".parse().unwrap(),
-        ));
-
-        let provider = ProviderBuilder::new()
-            .network::<Ethereum>()
-            .connect_http("http://localhost:8545".parse().unwrap());
-
-        let l2oo_contract = OPSuccinctL2OOContract::new(
-            address!("0xDafA1019F21AB8B27b319B1085f93673F02A69B7"),
-            provider.clone(),
-        );
-
-        let latest_header = provider.get_block(BlockId::latest()).await.unwrap().unwrap();
-
-        let transaction_request = l2oo_contract
-            .checkpointBlockHash(U256::from(latest_header.header.number))
-            .into_transaction_request();
-
-        let receipt = proposer_signer
-            .send_transaction_request("http://localhost:8545".parse().unwrap(), transaction_request)
-            .await
-            .unwrap();
-
-        println!("Signed transaction receipt: {receipt:?}");
-    }
-
-    #[tokio::test]
-    #[ignore]
-    // This test is meant to be ran locally to test various signers implementations,
-    // depending of the envvars set.
-    async fn test_send_transaction_request() {
-        dotenv::dotenv().ok();
-
-        rustls::crypto::aws_lc_rs::default_provider()
-            .install_default()
-            .expect("Failed to install default crypto provider");
-        let signer = SignerLock::from_env().await.unwrap();
-
-        println!("Signer: {}", signer.address());
-
-        let transaction_request = TransactionRequest::default()
-            .to(Address::from([
-                1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20,
-            ]))
-            .value(U256::from(100000u64))
-            .from(signer.address());
-        let receipt = signer
-            .send_transaction_request("http://localhost:8545".parse().unwrap(), transaction_request)
-            .await
-            .unwrap();
-        println!("Signed transaction receipt: {receipt:?}");
-    }
-}
